backend/requirements.txt
# Koru Backend Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Configuration
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# AI/ML Providers
openai>=1.0.0
google-genai>=1.0.0
elevenlabs>=1.0.0

# Async file operations
aiofiles>=23.0.0

# Multipart form data (for file uploads)
python-multipart>=0.0.6

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
